Make timestamp plugin available to all the adapters
