+++ lib/hanami/model/plugins.rb
@@ -1,0 +1,13 @@
+module Hanami
+  module Model
+    module Plugins
+      class WrappingInput
+        def initialize(relation, input)
+          @input = input || Hash
+        end
+      end
+
+      require 'hanami/model/plugins/timestamps'
+    end
+  end
+end
